+++ docs/reference/host-core-reset.md
@@ -138,6 +138,8 @@ The current native-host Swift split is:
   release-watchdog scheduling for AppKit interactions whose mouse-up event can be missed.
 - `CaptureHostPointerDispatch.swift`: capture-host pointer dispatch events, a shared delivery queue
   with per-track throttling state, and AppKit-to-controller pointer delivery support.
+- `CaptureHostView+LivePrimaryInteraction.swift`: capture-host live primary interaction release
+  recovery, pointer-preview mutation, mouse-up monitor wiring, and release-watchdog orchestration.
 - `LiveOverlayRenderer.swift`: live overlay render orchestration for HUD, loupe, frozen-pending
   preview, and live chrome layers.
 - `LiveHudColorRollCoordinator.swift`: live HUD color swatch/hex presentation, pending color roll
